% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.ihydro}
\alias{as.ihydro}
\title{ihydro Class}
\usage{
as.ihydro(file)
}
\arguments{
\item{file}{filepath to a geopackage file (.gpkg)}
}
\value{
an object of class ihydro
}
\description{
Creates an object of class "ihydro".
}
