% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrib_points.R
\name{attrib_points}
\alias{attrib_points}
\title{Title}
\usage{
attrib_points(
  input,
  spec = NULL,
  all_reaches = F,
  weighting_scheme = c("lumped", "iEucO", "iEucS", "iFLO", "iFLS", "HAiFLO", "HAiFLS"),
  loi_numeric_stats = c("distwtd_mean", "distwtd_sd", "mean", "sd", "median", "min",
    "max", "sum", "cell_count"),
  OS_combine = F,
  target_streamseg = F,
  buffer = 30,
  inv_function = function(x) {
     (x * 0.001 + 1)^-1
 },
  return_products = F,
  remove_region = NULL,
  temp_dir = NULL,
  verbose = F
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
