% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttrib_points.R
\name{.attr_fn}
\alias{.attr_fn}
\title{Calculate loi attributes from input rasters}
\usage{
.attr_fn(
  df = NULL,
  db_path = NULL,
  tbl_name = NULL,
  point_id,
  weighting_scheme2,
  loi_meta2,
  loi_cols2,
  loi_numeric_stats2,
  backend = c("data.table", "SQLite")
)
}
\value{
a data.frame of attributes
}
\description{
Calculate loi attributes from input rasters
}
\keyword{internal}
