% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttrib_points.R
\name{.summ_fn}
\alias{.summ_fn}
\title{Process gpkg rasters for attribute calculation}
\usage{
.summ_fn(
  x,
  input,
  iDW_file,
  loi_file,
  loi_meta,
  weighting_scheme,
  loi_numeric_stats,
  loi_cols,
  p,
  temp_dir_sub,
  n_cores
)
}
\value{
a data.frame of attributes
}
\description{
Process gpkg rasters for attribute calculation
}
\keyword{internal}
