% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vectors.R
\name{generate_vectors}
\alias{generate_vectors}
\title{Generate polygon subbasins, and attributed stream lines and points}
\usage{
generate_vectors(
  input,
  extra_attr = c("link_slope", "cont_slope", "USChnLn_To", "Elevation", "StOrd_Hack",
    "StOrd_Str", "StOrd_Hort", "StOrd_Shr"),
  return_products = F,
  temp_dir = NULL,
  verbose = F
)
}
\arguments{
\item{input}{resulting object from \code{generate_subbasins()}}

\item{extra_attr}{character. One or more of c("link_slope", "cont_slope", "USChnLn_To", "Elevation", "StOrd_Hack", "StOrd_Str", "StOrd_Hort", "StOrd_Shr"). Optional attributes to add to stream vector outputs.}

\item{return_products}{logical. If \code{TRUE}, a list containing the file path to write resulting \code{*.zip} file, and resulting GIS products. If \code{FALSE}, file path only.}

\item{temp_dir}{character. File path for temporary file storage, If \code{NULL}, \code{tempfile()} will be used}

\item{verbose}{logical.}
}
\value{
If \code{return_products = TRUE}, all geospatial analysis products are returned. If \code{return_products = FALSE}, folder path to resulting .zip file.
}
\description{
Generate polygon subbasins, and attributed stream lines and points
}
