% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ihydro_layers}
\alias{ihydro_layers}
\title{Describe layers objects of class ihydro}
\usage{
ihydro_layers(input)
}
\arguments{
\item{input}{an object of class ihydro}
}
\value{
table of objects in the ihydro geopackage
}
\description{
Describe layers objects of class ihydro
}
