% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_points.R
\name{insert_points}
\alias{insert_points}
\title{Title}
\usage{
insert_points(
  input,
  points,
  site_id_col,
  snap_distance,
  return_products = F,
  temp_dir = NULL,
  verbose = F
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
