% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_weights.R
\name{prep_weights}
\alias{prep_weights}
\title{Prepares hydroweights and adds them to existing zip file}
\usage{
prep_weights(
  input,
  sample_points = NULL,
  link_id = NULL,
  target_o_type = c("point", "segment_point", "segment_whole"),
  weighting_scheme = c("iEucS", "iFLS", "HAiFLS", "iEucO", "iFLO", "HAiFLO"),
  inv_function = function(x) {
     (x * 0.001 + 1)^-1
 },
  temp_dir = NULL,
  verbose = F
)
}
\arguments{
\item{input}{output from \code{process_hydrology()} (if \code{process_loi()} was not run on \code{process_hydrology()}, \code{loi_file} must be specified)}

\item{sample_points}{character or NULL. IDs of unique station identifiers priveded in 'site_id_col' of \code{generate_vectors()}}

\item{link_id}{character or NULL. 'link_id's of reaches to calculate attributes for.}

\item{target_o_type}{character. One of: c("point","segment_point","segment_whole"). Target for iEucO" "iFLO", and "HAiFLO" weighting schemes. "Point" represents the sampling point on the stream, "segment_point" represents the upstream segment of the sampling points, and "segment_whole" will target the entire reach, regardless of where sampling occurred.}

\item{weighting_scheme}{character. One or more weighting schemes: c("lumped", "iEucO", "iEucS", "iFLO", "iFLS", "HAiFLO", "HAiFLS")}

\item{inv_function}{function or named list of functions based on \code{weighting_scheme} names. Inverse function used in \code{terra::app()} to convert distances to inverse distances. Default: \code{(X * 0.001 + 1)^-1} assumes projection is in distance units of m and converts to distance units of km.}

\item{temp_dir}{character. File path for intermediate products; these are deleted once the function runs successfully.}

\item{verbose}{logical.}
}
\value{
input
}
\description{
Prepares hydroweights and adds them to existing zip file
}
