% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hydrology.R
\name{process_hydrology}
\alias{process_hydrology}
\title{Generate geospatial analysis products}
\usage{
process_hydrology(
  dem,
  output_filename,
  threshold,
  burn_streams = NULL,
  burn_depth = 5,
  min_length = NULL,
  depression_corr = c(NULL, "fill", "breach"),
  calc_catch = c("all", "none", "sample_points"),
  extra_attr = c("link_slope", "cont_slope", "USChnLn_To", "Elevation", "StOrd_Hack",
    "StOrd_Str", "StOrd_Hort", "StOrd_Shr"),
  points = NULL,
  snap_distance = NULL,
  break_on_noSnap = T,
  site_id_col = NULL,
  return_products = F,
  temp_dir = NULL,
  compress = F,
  verbose = F
)
}
\arguments{
\item{dem}{character (full file path with extension, e.g., "C:/Users/Administrator/Desktop/dem.tif"), \code{RasterLayer}, \code{SpatRaster}, or \code{PackedSpatRaster} of GeoTiFF type. Digital elevation model raster.}

\item{output_filename}{character. File path to write resulting .zip file.}

\item{threshold}{integer. Flow accumulation threshold for stream initiation.}

\item{burn_streams}{character. (full file path with extension, e.g., "C:/Users/Administrator/Desktop/input.shp"), sf, SpatVector, PackedSpatVector, RasterLayer, SpatRaster, or PackedSpatRaster. Stream vector to burn into DEM.}

\item{burn_depth}{numeric. Depth (in meters) to burn stream into the DEM.}

\item{min_length}{numeric. Minimum tributary length, shorter 1st order tributaries are removed.}

\item{depression_corr}{NULL or character. One of c("fill","breach"), specifying whether depressions should be filled or breached. NULL will perform neither, if DEM is already corrected.}

\item{extra_attr}{character. One or more of c("link_slope", "cont_slope", "USChnLn_To", "Elevation", "StOrd_Hack", "StOrd_Str", "StOrd_Hort", "StOrd_Shr"). Optional attributes to add to stream vector outputs.}

\item{points}{character (full file path with extension, e.g., "C:/Users/Administrator/Desktop/points.shp"), or any GIS data object that will be converted to spatial points. Points representing sampling locations.}

\item{snap_distance}{integer. Maximum distance which points will be snapped to stream lines in map units}

\item{site_id_col}{character. Variable name in \code{points} that corresponds to unique site identifiers. This column will be included in all vector geospatial analysis products. Note, if multiple points have the same \code{site_id_col}, their centroid will be used and returned; if multiple points overlap after snapping, only the first is used.}

\item{return_products}{logical. If \code{TRUE}, a list containing all geospatial analysis products. If \code{FALSE}, folder path to resulting .zip file.}

\item{temp_dir}{character. File path for intermediate products; these are deleted once the function runs successfully.}

\item{compress}{logical. Should output rasters be compressed, slower but more space efficient.}

\item{verbose}{logical.}
}
\value{
If \code{return_products = TRUE}, all geospatial analysis products are returned. If \code{return_products = FALSE}, folder path to resulting .zip file.
}
\description{
\code{ihydro::process_hydrology} processes a DEM into various geospatial analysis products.
}
