% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_loi.R
\name{process_loi}
\alias{process_loi}
\title{Title}
\usage{
process_loi(
  input = NULL,
  dem = NULL,
  clip_region = NULL,
  numb_inputs = list(),
  cat_inputs = list(),
  output_filename = NULL,
  variable_names = NULL,
  return_products = F,
  temp_dir = NULL,
  verbose = F
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
