% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrib_points.R
\name{attrib_points}
\alias{attrib_points}
\title{Attributes stream segments/sampling points with layers of interest (loi)}
\usage{
attrib_points(
  input,
  out_filename,
  loi_file = NULL,
  loi_cols = NULL,
  sample_points = NULL,
  link_id = NULL,
  clip_region = NULL,
  OS_combine = F,
  target_o_type = c("point", "segment_point", "segment_whole"),
  weighting_scheme = c("lumped", "iEucO", "iEucS", "iFLO", "iFLS", "HAiFLO", "HAiFLS"),
  loi_numeric_stats = c("distwtd_mean", "distwtd_sd", "mean", "sd", "median", "min",
    "max", "sum", "cell_count"),
  inv_function = function(x) {
     (x * 0.001 + 1)^-1
 },
  temp_dir = NULL,
  return_products = T,
  verbose = F
)
}
\arguments{
\item{input}{an object of class ihydro}

\item{out_filename}{filepath of csv to save resulting attributes}

\item{loi_file}{filepath of \code{process_loi()} output}

\item{loi_cols}{character or NULL. Names of loi layers to include in summary. If NULL, all layers used.}

\item{sample_points}{character or NULL. IDs of unique station identifiers provided in 'site_id_col' of \code{generate_vectors()} to calculate attributes for.}

\item{link_id}{character or NULL. 'link_id's of reaches to calculate attributes for.}

\item{clip_region}{character (full file path with extension, e.g., "C:/Users/Administrator/Desktop/lu.shp"), \code{sf}, \code{SpatVector}, \code{PackedSpatVector}, \code{RasterLayer}, \code{SpatRaster}, or \code{PackedSpatRaster}. Regions to remove when summarizing the attributes (e.g., remove lake from catchment)}

\item{OS_combine}{logical. Should target_O and target_S be merged as targets for iEucS, iFLS, and/or HAiFLS? Use \code{TRUE} or \code{FALSE}. This allows cells surrounding \code{target_O} to flow directly into \code{target_O} rather than be forced through \code{target_S}.}

\item{target_o_type}{character. One of: c("point","segment_point","segment_whole"). Target for "iEucO", "iFLO", and "HAiFLO" weighting schemes. "Point" represents the sampling point on the stream, "segment_point" represents the upstream segment of the sampling points, and "segment_whole" will target the entire reach, regardless of where sampling occurred.}

\item{weighting_scheme}{character. One or more weighting schemes: c("lumped", "iEucO", "iEucS", "iFLO", "iFLS", "HAiFLO", "HAiFLS")}

\item{loi_numeric_stats}{character. One or more of c("distwtd_mean", "distwtd_sd", "mean", "sd", "median", "min", "max", "sum", "cell_count"). Those without distwtd_ are simple "lumped" statistics.}

\item{inv_function}{function or named list of functions based on \code{weighting_scheme} names. Inverse function used in \code{terra::app()} to convert distances to inverse distances. Default: \code{(X * 0.001 + 1)^-1} assumes projection is in distance units of m and converts to distance units of km.}

\item{temp_dir}{character. File path for intermediate products; these are deleted once the function runs successfully.}

\item{return_products}{logical. If \code{TRUE}, a list containing all geospatial analysis products. If \code{FALSE}, folder path to resulting .zip file.}

\item{verbose}{logical.}
}
\value{
A tibble of resulting attributes. If \code{return_products = TRUE}, all geospatial analysis products are returned as well.
}
\description{
Attributes stream segments/sampling points with layers of interest (loi)
}
