% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrib_streamline.R
\name{attrib_streamline}
\alias{attrib_streamline}
\title{Title}
\usage{
attrib_streamline(
  input,
  extra_attr = c("link_slope", "cont_slope", "USChnLn_To", "Elevation", "StOrd_Hack",
    "StOrd_Str", "StOrd_Hort", "StOrd_Shr"),
  return_products = F,
  temp_dir = NULL,
  verbose = F
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
