% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pwisedist.R
\name{generate_pwisedist}
\alias{generate_pwisedist}
\title{Generate pairwise distances between flow-connected and flow-unconnected site pairs, as well proportions of shared catchments}
\usage{
generate_pwisedist(input, return_products = F, temp_dir = NULL, verbose = F)
}
\arguments{
\item{input}{ouptut of \code{trace_flowpaths()}}

\item{return_products}{logical. If \code{TRUE}, a list containing the file path to write resulting \code{*.zip} file, and resulting GIS products. If \code{FALSE}, file path only.}

\item{temp_dir}{character. File path for temporary file storage, If \code{NULL}, \code{tempfile()} will be used}

\item{verbose}{logical.}
}
\value{
If \code{return_products = TRUE}, all geospatial analysis products are returned. If \code{return_products = FALSE}, folder path to resulting .zip file.
}
\description{
Generate pairwise distances between flow-connected and flow-unconnected site pairs, as well proportions of shared catchments
}
