% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_subbasins.R
\name{generate_subbasins}
\alias{generate_subbasins}
\title{Extracts polygon subbasins from 'process_flowdir()'}
\usage{
generate_subbasins(input, return_products = F, temp_dir = NULL, verbose = F)
}
\arguments{
\item{input}{resulting object from \code{process_flowdir()}}

\item{return_products}{logical. If \code{TRUE}, a list containing the file path to write resulting \code{*.zip} file, and resulting GIS products. If \code{FALSE}, file path only.}

\item{temp_dir}{character. File path for temporary file storage, If \code{NULL}, \code{tempfile()} will be used}

\item{verbose}{logical.}
}
\value{
If \code{return_products = TRUE}, all geospatial analysis products are returned. If \code{return_products = FALSE}, folder path to resulting .zip file.
}
\description{
Extracts polygon subbasins from 'process_flowdir()'
}
\seealso{
\link[whitebox:wbt_subbasins]{whitebox::wbt_subbasins}
}
