% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hydrology.R
\name{process_hydrology}
\alias{process_hydrology}
\title{Process DEM by generating various terrain products}
\usage{
process_hydrology(
  dem,
  output_filename,
  threshold,
  extra_attr = c("link_slope", "cont_slope", "USChnLn_To", "Elevation", "StOrd_Hack",
    "StOrd_Str", "StOrd_Hort", "StOrd_Shr"),
  points = NULL,
  snap_distance = NULL,
  site_id_col = NULL,
  return_products = F,
  temp_dir = NULL,
  verbose = F
)
}
\arguments{
\item{dem}{character (full file path with extension, e.g., "C:/Users/Administrator/Desktop/dem.tif"), \code{RasterLayer}, \code{SpatRaster}, or \code{PackedSpatRaster} of GeoTiFF type. Digital elevation model raster.}

\item{output_filename}{character. File path to write resulting .zip file.}

\item{threshold}{numeric. Threshold in flow accumulation values for channelization.}

\item{return_products}{logical. If \code{TRUE}, a list containing all geospatial analysis products. If \code{FALSE}, folder path to resulting .zip file.}

\item{temp_dir}{character. File path for intermediate products; these are deleted once the function runs successfully.}

\item{verbose}{logical. If \code{FALSE}, the function will not print output prints.}
}
\value{
If \code{return_products = TRUE}, all geospatial analysis products are returned. If \code{return_products = FALSE}, folder path to resulting .zip file.
}
\description{
#' \code{IntegratedHydrology::process_hydrology()} processes a DEM into various geospatial analysis products.
}
\examples{

library(whitebox)
library(terra)
library(sf)
library(hydroweight)

## Import toy_dem from whitebox package
toy_file<-sample_dem_data()
toy_file <- system.file("extdata", "DEM.tif", package = "whitebox")
toy_dem <- rast(raster::raster(x = toy_file)) # reading the file from terra directly sometimes crashes R for some reason
crs(toy_dem) <- "epsg:3161"

## Generate hydroweight_dir as a temporary directory
save_dir <- tempdir()

hydro_out<-process_hydrology(
  dem=toy_dem,
  threshold=1000,
  return_products=T,
  save_dir=save_dir,
  temp_dir=NULL,
  verbose=F
)

}
